break fortune_teller into typical ruby project structure
