+++ README.md
@@ -4,4 +4,4 @@ clone to your machine
 
 cd to the folder in your terminal 
 
-summon the digital fortune teller by typing 'ruby fortuneteller.rb'
+summon the digital fortune teller by typing `bin/fortune_teller`
