+++ lib/prediction_generator.rb
