+++ lib/slow_type.rb
@@ -2,7 +2,7 @@ module SlowType
   def self.creep_type(text)
     text = text.split('')
     text.each do |l|
-      sleep 0.08
+      # sleep 0.08
       print l
     end
     print "\n"
